requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.6.0
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
phonenumbers==8.12.47
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
